+++ src/app/app.module.ts
@@ -1,3 +1,4 @@
+import { TagsService } from './services/tags.service';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
@@ -53,7 +54,7 @@ import { ProjectDataService } from './services/project-data.service';
       enabled: environment.production
     })
   ],
-  providers: [ProjectDataService],
+  providers: [ProjectDataService, TagsService],
   bootstrap: [AppComponent]
 })
 export class AppModule {}
