+++ src/app/utils/ITags.ts
@@ -1,0 +1,4 @@
+export interface ITag {
+  displayName: string;
+  isSelected: boolean;
+}
