+++ src/app/utils/tags-data.ts
@@ -1,0 +1,7 @@
+export const tagData = [
+  { displayName: 'Angular', isSelected: true },
+  { displayName: 'React.js', isSelected: true },
+  { displayName: 'Node.js', isSelected: true },
+  { displayName: 'ASP.NET', isSelected: true },
+  { displayName: 'Others', isSelected: !true }
+];
